(feat) record and report console logs and network events.
this commit also refactors the internal logic a little. the most
significant being that both a preflight and screenshot run needs
to be completed before any data will be persisted.

the function that actually took screenshots also got a signature
change to cater for a more complex return type, should we wish
to capture more / othere data in the future.